TensorFlow.Minimize

Documentation

type Minimizer a = forall m. MonadBuild m => [Variable a] -> [Tensor Value a] -> m ControlNode

Functions that minimize a loss w.r.t. a set of Variables.

Generally only performs one step of an iterative algorithm.

Minimizers are defined as a function of the gradients instead of the loss so that users can apply transformations to the gradients.

minimizeWith

Arguments

:: (MonadBuild m, GradientCompatible a)

=> Minimizer a

-> Tensor v a
Loss.

-> [Variable a]
Parameters of the loss function.

-> m ControlNode

Convenience wrapper around gradients and a Minimizer.

gradientDescent

Arguments

:: GradientCompatible a

=> a
Learning rate.

-> Minimizer a

Perform one step of the gradient descent algorithm.

type OneOfAdamDataTypes t = OneOf '[Complex Double, Complex Float, Int16, Int32, Int64, Int8, Word16, Word32, Word64, Word8, Double, Float] t

data AdamConfig t

Constructors

AdamConfig

Fields

adamLearningRate :: t

adamBeta1 :: t

adamBeta2 :: t

adamEpsilon :: t

adam :: (OneOfAdamDataTypes t, Fractional t) => Minimizer t

Perform one step of the adam algorithm.

See https://arxiv.org/abs/1412.6980.

NOTE: Currently requires all Variables to have an initializedValue.

adam' :: OneOfAdamDataTypes t => AdamConfig t -> Minimizer t