TensorFlow.Minimize

Documentation

type Minimizer a = forall m. MonadBuild m => [Variable a] -> [Tensor Value a] -> m ControlNode

Functions that minimize a loss w.r.t. a set of Variables.

Generally only performs one step of an iterative algorithm.

Minimizers are defined as a function of the gradients instead of the loss so that users can apply transformations to the gradients.

minimizeWith

Arguments

:: (MonadBuild m, GradientCompatible a)

=> Minimizer a

-> Tensor v a
Loss.

-> [Variable a]
Parameters of the loss function.

-> m ControlNode

Convenience wrapper around gradients and a Minimizer.

gradientDescent

Arguments

:: GradientCompatible a

=> a
Learning rate.

-> Minimizer a

Perform one step of the gradient descent algorithm.

data AdamConfig

Constructors

AdamConfig

Fields

adamLearningRate :: Float

adamBeta1 :: Float

adamBeta2 :: Float

adamEpsilon :: Float

adam :: Minimizer Float

Perform one step of the adam algorithm.

See https://arxiv.org/abs/1412.6980.

NOTE: Currently requires all Variables to have an initializedValue.

adam' :: AdamConfig -> Minimizer Float